+++ plugins/source/aws/resources/services/iam/group_policies.go
@@ -1,7 +1,13 @@
 package iam
 
 import (
+	"context"
+	"encoding/json"
+	"net/url"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/iam"
+	"github.com/aws/aws-sdk-go-v2/service/iam/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/schema"
 	"github.com/cloudquery/plugin-sdk/transformers"
@@ -35,4 +41,60 @@ func groupPolicies() *schema.Table {
 			},
 		},
 	}
+}
+
+func fetchIamGroupPolicies(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	c := meta.(*client.Client)
+	svc := c.Services().Iam
+	group := parent.Item.(types.Group)
+	config := iam.ListGroupPoliciesInput{
+		GroupName: group.GroupName,
+	}
+	for {
+		output, err := svc.ListGroupPolicies(ctx, &config)
+		if err != nil {
+			if c.IsNotFoundError(err) {
+				return nil
+			}
+			return err
+		}
+
+		res <- output.PolicyNames
+
+		if aws.ToString(output.Marker) == "" {
+			break
+		}
+		config.Marker = output.Marker
+	}
+	return nil
+}
+
+func getGroupPolicy(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource) error {
+	c := meta.(*client.Client)
+	svc := c.Services().Iam
+	p := resource.Item.(string)
+	group := resource.Parent.Item.(types.Group)
+
+	policyResult, err := svc.GetGroupPolicy(ctx, &iam.GetGroupPolicyInput{PolicyName: &p, GroupName: group.GroupName})
+	if err != nil {
+		return err
+	}
+	resource.Item = policyResult
+	return nil
+}
+
+func resolveIamGroupPolicyPolicyDocument(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
+	r := resource.Item.(*iam.GetGroupPolicyOutput)
+
+	decodedDocument, err := url.QueryUnescape(*r.PolicyDocument)
+	if err != nil {
+		return err
+	}
+
+	var document map[string]any
+	err = json.Unmarshal([]byte(decodedDocument), &document)
+	if err != nil {
+		return err
+	}
+	return resource.Set(c.Name, document)
 }
