--- plugins/source/aws/resources/services/iam/groups_fetch.go
@@ -1,1 +1,0 @@
-package iam
