+++ plugins/source/aws/resources/services/iam/instance_profiles.go
@@ -1,6 +1,9 @@
 package iam
 
 import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/service/iam"
 	"github.com/aws/aws-sdk-go-v2/service/iam/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/schema"
@@ -32,4 +35,32 @@ func InstanceProfiles() *schema.Table {
 			},
 		},
 	}
+}
+
+func fetchIamInstanceProfiles(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	config := iam.ListInstanceProfilesInput{}
+	svc := meta.(*client.Client).Services().Iam
+	p := iam.NewListInstanceProfilesPaginator(svc, &config)
+	for p.HasMorePages() {
+		response, err := p.NextPage(ctx)
+		if err != nil {
+			return err
+		}
+		res <- response.InstanceProfiles
+	}
+	return nil
+}
+
+func resolveIamInstanceProfileTags(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
+	r := resource.Item.(types.InstanceProfile)
+	cl := meta.(*client.Client)
+	svc := cl.Services().Iam
+	response, err := svc.ListInstanceProfileTags(ctx, &iam.ListInstanceProfileTagsInput{InstanceProfileName: r.InstanceProfileName})
+	if err != nil {
+		if cl.IsNotFoundError(err) {
+			return nil
+		}
+		return err
+	}
+	return resource.Set("tags", client.TagsToMap(response.Tags))
 }
