+++ plugins/source/aws/resources/services/iam/roles.go
@@ -1,6 +1,11 @@
 package iam
 
 import (
+	"context"
+	"encoding/json"
+	"net/url"
+
+	"github.com/aws/aws-sdk-go-v2/service/iam"
 	"github.com/aws/aws-sdk-go-v2/service/iam/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/schema"
@@ -44,8 +49,76 @@ func Roles() *schema.Table {
 		},
 
 		Relations: []*schema.Table{
-			RolePolicies(),
+			rolePolicies(),
 			roleLastAccessedDetails(),
 		},
 	}
+}
+
+func fetchIamRoles(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	var config iam.ListRolesInput
+	svc := meta.(*client.Client).Services().Iam
+	paginator := iam.NewListRolesPaginator(svc, &config)
+	for paginator.HasMorePages() {
+		response, err := paginator.NextPage(ctx)
+		if err != nil {
+			return err
+		}
+		res <- response.Roles
+	}
+	return nil
+}
+
+func getRole(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource) error {
+	role := resource.Item.(types.Role)
+	svc := meta.(*client.Client).Services().Iam
+	roleDetails, err := svc.GetRole(ctx, &iam.GetRoleInput{
+		RoleName: role.RoleName,
+	})
+	if err != nil {
+		return err
+	}
+	resource.Item = roleDetails.Role
+	return nil
+}
+
+func resolveIamRolePolicies(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
+	r := resource.Item.(*types.Role)
+	cl := meta.(*client.Client)
+	svc := cl.Services().Iam
+	input := iam.ListAttachedRolePoliciesInput{
+		RoleName: r.RoleName,
+	}
+	policies := map[string]*string{}
+	paginator := iam.NewListAttachedRolePoliciesPaginator(svc, &input)
+	for paginator.HasMorePages() {
+		response, err := paginator.NextPage(ctx)
+		if err != nil {
+			if cl.IsNotFoundError(err) {
+				return nil
+			}
+			return err
+		}
+		for _, p := range response.AttachedPolicies {
+			policies[*p.PolicyArn] = p.PolicyName
+		}
+	}
+	return resource.Set("policies", policies)
+}
+
+func resolveRolesAssumeRolePolicyDocument(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
+	r := resource.Item.(*types.Role)
+	if r.AssumeRolePolicyDocument == nil {
+		return nil
+	}
+	decodedDocument, err := url.QueryUnescape(*r.AssumeRolePolicyDocument)
+	if err != nil {
+		return err
+	}
+	var d map[string]any
+	err = json.Unmarshal([]byte(decodedDocument), &d)
+	if err != nil {
+		return err
+	}
+	return resource.Set("assume_role_policy_document", d)
 }
