--- plugins/source/aws/resources/services/iam/server_certificates_fetch.go
@@ -1,23 +1,0 @@
-package iam
-
-import (
-	"context"
-
-	"github.com/aws/aws-sdk-go-v2/service/iam"
-	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
-	"github.com/cloudquery/plugin-sdk/schema"
-)
-
-func fetchIamServerCertificates(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
-	var config iam.ListServerCertificatesInput
-	svc := meta.(*client.Client).Services().Iam
-	paginator := iam.NewListServerCertificatesPaginator(svc, &config)
-	for paginator.HasMorePages() {
-		page, err := paginator.NextPage(ctx)
-		if err != nil {
-			return err
-		}
-		res <- page.ServerCertificateMetadataList
-	}
-	return nil
-}
