+++ plugins/source/aws/resources/services/iam/signing_certificates.go
@@ -1,13 +1,16 @@
 package iam
 
 import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/service/iam"
 	"github.com/aws/aws-sdk-go-v2/service/iam/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/schema"
 	"github.com/cloudquery/plugin-sdk/transformers"
 )
 
-func SigningCertificates() *schema.Table {
+func signingCertificates() *schema.Table {
 	tableName := "aws_iam_signing_certificates"
 	return &schema.Table{
 		Name:        tableName,
@@ -32,4 +35,18 @@ func SigningCertificates() *schema.Table {
 			},
 		},
 	}
+}
+
+func fetchUserSigningCertificates(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	config := &iam.ListSigningCertificatesInput{UserName: parent.Item.(*types.User).UserName}
+	svc := meta.(*client.Client).Services().Iam
+	paginator := iam.NewListSigningCertificatesPaginator(svc, config)
+	for paginator.HasMorePages() {
+		page, err := paginator.NextPage(ctx)
+		if err != nil {
+			return err
+		}
+		res <- page.Certificates
+	}
+	return nil
 }
