+++ plugins/source/aws/resources/services/iam/signing_certificates_fetch.go
@@ -1,24 +1,1 @@
 package iam
-
-import (
-	"context"
-
-	"github.com/aws/aws-sdk-go-v2/service/iam"
-	"github.com/aws/aws-sdk-go-v2/service/iam/types"
-	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
-	"github.com/cloudquery/plugin-sdk/schema"
-)
-
-func fetchUserSigningCertificates(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
-	config := &iam.ListSigningCertificatesInput{UserName: parent.Item.(*types.User).UserName}
-	svc := meta.(*client.Client).Services().Iam
-	paginator := iam.NewListSigningCertificatesPaginator(svc, config)
-	for paginator.HasMorePages() {
-		page, err := paginator.NextPage(ctx)
-		if err != nil {
-			return err
-		}
-		res <- page.Certificates
-	}
-	return nil
-}
