+++ plugins/source/aws/resources/services/iam/ssh_public_keys.go
@@ -1,13 +1,16 @@
 package iam
 
 import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/service/iam"
 	"github.com/aws/aws-sdk-go-v2/service/iam/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/schema"
 	"github.com/cloudquery/plugin-sdk/transformers"
 )
 
-func SshPublicKeys() *schema.Table {
+func sshPublicKeys() *schema.Table {
 	tableName := "aws_iam_ssh_public_keys"
 	return &schema.Table{
 		Name:        tableName,
@@ -37,4 +40,18 @@ func SshPublicKeys() *schema.Table {
 			},
 		},
 	}
+}
+func fetchIamSshPublicKeys(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	svc := meta.(*client.Client).Services().Iam
+	paginator := iam.NewListSSHPublicKeysPaginator(svc, &iam.ListSSHPublicKeysInput{
+		UserName: parent.Item.(*types.User).UserName,
+	})
+	for paginator.HasMorePages() {
+		page, err := paginator.NextPage(ctx)
+		if err != nil {
+			return err
+		}
+		res <- page.SSHPublicKeys
+	}
+	return nil
 }
