--- plugins/source/aws/resources/services/iam/user_policies_fetch.go
@@ -1,66 +1,0 @@
-package iam
-
-import (
-	"context"
-	"encoding/json"
-	"net/url"
-
-	"github.com/aws/aws-sdk-go-v2/aws"
-	"github.com/aws/aws-sdk-go-v2/service/iam"
-	"github.com/aws/aws-sdk-go-v2/service/iam/types"
-	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
-	"github.com/cloudquery/plugin-sdk/schema"
-)
-
-func fetchIamUserPolicies(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
-	c := meta.(*client.Client)
-	svc := c.Services().Iam
-	user := parent.Item.(*types.User)
-	config := iam.ListUserPoliciesInput{UserName: user.UserName}
-	for {
-		output, err := svc.ListUserPolicies(ctx, &config)
-		if err != nil {
-			if c.IsNotFoundError(err) {
-				return nil
-			}
-			return err
-		}
-		res <- output.PolicyNames
-
-		if aws.ToString(output.Marker) == "" {
-			break
-		}
-		config.Marker = output.Marker
-	}
-	return nil
-}
-
-func getUserPolicy(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource) error {
-	svc := meta.(*client.Client).Services().Iam
-	p := resource.Item.(string)
-	user := resource.Parent.Item.(*types.User)
-
-	policyResult, err := svc.GetUserPolicy(ctx, &iam.GetUserPolicyInput{PolicyName: &p, UserName: user.UserName})
-	if err != nil {
-		return err
-	}
-
-	resource.Item = policyResult
-	return nil
-}
-
-func resolveIamUserPolicyPolicyDocument(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
-	r := resource.Item.(*iam.GetUserPolicyOutput)
-
-	decodedDocument, err := url.QueryUnescape(*r.PolicyDocument)
-	if err != nil {
-		return err
-	}
-
-	var document map[string]any
-	err = json.Unmarshal([]byte(decodedDocument), &document)
-	if err != nil {
-		return err
-	}
-	return resource.Set(c.Name, document)
-}
