+++ plugins/source/aws/resources/services/iam/users.go
@@ -1,6 +1,10 @@
 package iam
 
 import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/iam"
 	"github.com/aws/aws-sdk-go-v2/service/iam/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/schema"
@@ -39,13 +43,41 @@ func Users() *schema.Table {
 		},
 
 		Relations: []*schema.Table{
-			UserAccessKeys(),
-			UserGroups(),
-			UserAttachedPolicies(),
-			UserPolicies(),
-			SshPublicKeys(),
-			SigningCertificates(),
+			userAccessKeys(),
+			userGroups(),
+			userAttachedPolicies(),
+			userPolicies(),
+			sshPublicKeys(),
+			signingCertificates(),
 			userLastAccessedDetails(),
 		},
 	}
+}
+
+func fetchIamUsers(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	config := iam.ListUsersInput{}
+	c := meta.(*client.Client)
+	svc := c.Services().Iam
+	p := iam.NewListUsersPaginator(svc, &config)
+	for p.HasMorePages() {
+		response, err := p.NextPage(ctx)
+		if err != nil {
+			return err
+		}
+		res <- response.Users
+	}
+	return nil
+}
+
+func getUser(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource) error {
+	listUser := resource.Item.(types.User)
+	svc := meta.(*client.Client).Services().Iam
+	userDetail, err := svc.GetUser(ctx, &iam.GetUserInput{
+		UserName: aws.String(*listUser.UserName),
+	})
+	if err != nil {
+		return err
+	}
+	resource.Item = userDetail.User
+	return nil
 }
