+++ plugins/source/aws/resources/services/iam/users_fetch.go
@@ -3,109 +3,11 @@ package iam
 import (
 	"context"
 
-	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/iam"
 	"github.com/aws/aws-sdk-go-v2/service/iam/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
-	"github.com/cloudquery/cloudquery/plugins/source/aws/resources/services/iam/models"
 	"github.com/cloudquery/plugin-sdk/schema"
 )
-
-func fetchIamUsers(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
-	config := iam.ListUsersInput{}
-	c := meta.(*client.Client)
-	svc := c.Services().Iam
-	p := iam.NewListUsersPaginator(svc, &config)
-	for p.HasMorePages() {
-		response, err := p.NextPage(ctx)
-		if err != nil {
-			return err
-		}
-		res <- response.Users
-	}
-	return nil
-}
-
-func getUser(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource) error {
-	listUser := resource.Item.(types.User)
-	svc := meta.(*client.Client).Services().Iam
-	userDetail, err := svc.GetUser(ctx, &iam.GetUserInput{
-		UserName: aws.String(*listUser.UserName),
-	})
-	if err != nil {
-		return err
-	}
-	resource.Item = userDetail.User
-	return nil
-}
-
-func fetchIamUserGroups(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
-	var config iam.ListGroupsForUserInput
-	p := parent.Item.(*types.User)
-	svc := meta.(*client.Client).Services().Iam
-	config.UserName = p.UserName
-	paginator := iam.NewListGroupsForUserPaginator(svc, &config)
-	for paginator.HasMorePages() {
-		output, err := paginator.NextPage(ctx)
-		if err != nil {
-			return err
-		}
-		res <- output.Groups
-	}
-	return nil
-}
-
-func fetchIamUserAccessKeys(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
-	var config iam.ListAccessKeysInput
-	p := parent.Item.(*types.User)
-	svc := meta.(*client.Client).Services().Iam
-	config.UserName = p.UserName
-	for {
-		output, err := svc.ListAccessKeys(ctx, &config)
-		if err != nil {
-			return err
-		}
-
-		keys := make([]models.AccessKeyWrapper, len(output.AccessKeyMetadata))
-		for i, key := range output.AccessKeyMetadata {
-			switch i {
-			case 0:
-				keys[i] = models.AccessKeyWrapper{AccessKeyMetadata: key, LastRotated: *key.CreateDate}
-			case 1:
-				keys[i] = models.AccessKeyWrapper{AccessKeyMetadata: key, LastRotated: *key.CreateDate}
-			default:
-				keys[i] = models.AccessKeyWrapper{AccessKeyMetadata: key}
-			}
-		}
-		res <- keys
-		if output.Marker == nil {
-			break
-		}
-		config.Marker = output.Marker
-	}
-	return nil
-}
-
-func postIamUserAccessKeyResolver(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource) error {
-	r := resource.Item.(models.AccessKeyWrapper)
-	if r.AccessKeyId == nil {
-		return nil
-	}
-	svc := meta.(*client.Client).Services().Iam
-	output, err := svc.GetAccessKeyLastUsed(ctx, &iam.GetAccessKeyLastUsedInput{AccessKeyId: r.AccessKeyId})
-	if err != nil {
-		return err
-	}
-	if output.AccessKeyLastUsed != nil {
-		if err := resource.Set("last_used", output.AccessKeyLastUsed.LastUsedDate); err != nil {
-			return err
-		}
-		if err := resource.Set("last_used_service_name", output.AccessKeyLastUsed.ServiceName); err != nil {
-			return err
-		}
-	}
-	return nil
-}
 
 func fetchIamUserAttachedPolicies(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
 	p := parent.Item.(*types.User)
