+++ plugins/source/aws/resources/services/iam/virtual_mfa_devices.go
@@ -1,6 +1,10 @@
 package iam
 
 import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/iam"
 	"github.com/aws/aws-sdk-go-v2/service/iam/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/schema"
@@ -31,4 +35,22 @@ func VirtualMfaDevices() *schema.Table {
 			},
 		},
 	}
+}
+
+func fetchIamVirtualMfaDevices(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	var config iam.ListVirtualMFADevicesInput
+	svc := meta.(*client.Client).Services().Iam
+	for {
+		response, err := svc.ListVirtualMFADevices(ctx, &config)
+		if err != nil {
+			return err
+		}
+		res <- response.VirtualMFADevices
+		if aws.ToString(response.Marker) == "" {
+			break
+		}
+		config.Marker = response.Marker
+	}
+
+	return nil
 }
