--- plugins/source/aws/resources/services/iam/virtual_mfa_devices_fetch.go
@@ -1,28 +1,0 @@
-package iam
-
-import (
-	"context"
-
-	"github.com/aws/aws-sdk-go-v2/aws"
-	"github.com/aws/aws-sdk-go-v2/service/iam"
-	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
-	"github.com/cloudquery/plugin-sdk/schema"
-)
-
-func fetchIamVirtualMfaDevices(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
-	var config iam.ListVirtualMFADevicesInput
-	svc := meta.(*client.Client).Services().Iam
-	for {
-		response, err := svc.ListVirtualMFADevices(ctx, &config)
-		if err != nil {
-			return err
-		}
-		res <- response.VirtualMFADevices
-		if aws.ToString(response.Marker) == "" {
-			break
-		}
-		config.Marker = response.Marker
-	}
-
-	return nil
-}
